Update README.md
I made a project development on the frontand side with Angular and on the backend side with C#. I can call the API related to products and category from my backend 'Etrade' project in my angular project and select the products on the basis of category and add them to the cart. I can select more than one product in the cart and delete these products. I will also add a login page to my project. I plan to add a page where I will show product details and a picture of the products.